rest-mqtt-master/requirements.txt
click==6.7
Flask==0.12.2
Flask-API==1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
Werkzeug==0.14.1
